+++ .dumirc.ts
@@ -2,8 +2,23 @@ import { defineConfig } from 'dumi';
 
 export default defineConfig({
   outputPath: 'docs-dist',
+  favicons: [
+    '/favicons/favicon.ico',
+    '/favicons/favicon-16x16.png',
+    '/favicons/favicon-32x32.png',
+    '/favicons/favicon-64x64.png',
+  ],
+  logo: '/logos/nav-logo-375.png',
+  analytics: {
+    baidu: 'df33e0123cd576892df8bc4ac64788cf',
+    ga_v2: 'G-SC8THD4FR9',
+  },
+  sitemap: {
+    hostname: 'https://banana-ui.com',
+  },
   themeConfig: {
     name: 'Banana UI',
+    hideNameOnHeader: true,
     footer: false,
     hero: {
       showCustomContent: true,
@@ -130,5 +145,6 @@ export default defineConfig({
         },
       ],
     },
+    prefersColor: { default: 'light', switch: false },
   },
 });