+++ docs/example/Modal/demos/overflow.tsx
@@ -4,13 +4,13 @@
  */
 
 import { Button, Message, Modal } from '@banana-ui/react';
-import React, { useState } from 'react';
+import { useState } from 'react';
 
 export default function Overflow() {
   const [visible, setVisible] = useState(false);
 
   return (
-    <div>
+    <>
       <Button type="primary" onClick={() => setVisible(true)}>
         Open Modal
       </Button>
@@ -30,6 +30,6 @@ export default function Overflow() {
           <span>Hi~ o(*￣▽￣*)ブ</span>
         </div>
       </Modal>
-    </div>
+    </>
   );
 }