+++ docs/example/Select/demos/basicUsage.tsx
@@ -3,7 +3,6 @@
  */
 
 import { Select, SelectOption } from '@banana-ui/react';
-import React from 'react';
 
 export default function BasicUsage() {
   return (